# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "AsOfModel": {
        "properties": {
          "requested": {
            "description": "What the caller asked for; `latest` when unspecified.",
            "title": "Requested",
            "type": "string"
          },
          "resolved": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The concrete vintage actually served.",
            "title": "Resolved"
          }
        },
        "required": [
          "requested",
          "resolved"
        ],
        "title": "AsOfModel",
        "type": "object"
      },
...
      "EnvelopeModel_WellDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WellDetail"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[WellDetail]",
        "type": "object"
      },
...
      "LinksModel": {
        "additionalProperties": true,
        "properties": {
          "explain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pre-built /v1/explain call for every handle in the body.",
            "title": "Explain"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Next page, or null at the end.",
            "title": "Next"
          },
          "self": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Canonical path of this response.",
            "title": "Self"
          }
        },
        "title": "LinksModel",
        "type": "object"
      },
...
      "MetaModel": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "$ref": "#/components/schemas/AsOfModel",
            "description": "Requested and resolved vintage (SB-04 \u00a72.5)."
          },
          "deprecations": {
            "description": "Deprecations in force for this operation.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Deprecations",
            "type": "array"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "JSON Pointer to glossary term_id (DIR-8).",
            "title": "Labels",
            "type": "object"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor for the next page; null at the end.",
            "title": "Next Cursor"
          },
          "request_id": {
            "description": "ULID for this request; echoed in every problem body.",
            "title": "Request Id",
            "type": "string"
          },
          "source_freshness": {
            "additionalProperties": true,
            "description": "Per-source retrieval vintage and state.",
            "title": "Source Freshness",
            "type": "object"
          },
          "warnings": {
            "description": "Structured warnings: code, detail, optional pointer.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "as_of"
        ],
        "title": "MetaModel",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "The wire shape of every failure (RFC 9457).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What went wrong on this request.",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field failures, each with a JSON Pointer.",
            "title": "Errors"
          },
          "instance": {
            "description": "Path of the request that failed.",
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "description": "ULID joining this response to the audit trail.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the error code.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Absolute URI of the error code; resolves at /v1/errors/{code}.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Static owner key for this deployment. /healthz and the frontend are open.",
        "in": "header",
        "name": "X-Glasswell-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Glass-box upstream analytics on public data. Every number this API serves carries a\nderivation handle in band, and `GET /v1/explain` walks that handle back to the\nchecksummed government file it came from.\n\nConventions that hold everywhere: responses are `data` / `meta` / `links`; collections\nare cursor-paginated with no offset parameter; failures are RFC 9457 problem documents\nwhose `type` resolves at `/v1/errors/{code}`; and `as_of` selects knowledge time, with\nthe resolved vintage reported back in `meta.as_of`.\n\nThis deployment is the North Dakota slice: wells, geometry, monthly production, the\nlineage spine, the conformance registry, quarantine and the glossary. Forecasts,\neconomics and other basins are not served.",
    "title": "glasswell",
    "version": "0.1.0",
    "x-glasswell-freeze": {
      "checked_by": "tests/contract/openapi_diff.py",
      "criterion": "S1",
      "frozen_on": "2026-08-20",
      "policy": "Additive change only. A path, operation, parameter, response field or enum value that is published here is published for the life of /v1; removing one, or making an optional request parameter required, is a /v2 event (blueprint \u00a73.6.1).",
      "status": "frozen",
      "surface": "v1"
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/wells/{api10}": {
      "get": {
        "description": "Header, resolved status, land unit and geometry references for one well, plus the lateral count and total lateral length. Length is computed live from canonical.well_spatial in the basin's compute CRS \u2014 never from a stored degree length, and never from a mart this slice cannot exercise. It does not include production, forecasts or economics.",
        "operationId": "get_well",
        "parameters": [
          {
            "description": "Ten-digit API well number.",
            "in": "path",
            "name": "api10",
            "required": true,
            "schema": {
              "description": "Ten-digit API well number.",
              "pattern": "^\\d{10}$",
              "title": "Api10",
              "type": "string"
            }
          },
          {
            "description": "Knowledge-time cut: serve the greatest vintage at or before this date. Defaults to the latest published vintage, never to wall-clock now.",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Knowledge-time cut: serve the greatest vintage at or before this date. Defaults to the latest published vintage, never to wall-clock now.",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_WellDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "not_found": {
                    "summary": "Not found",
                    "value": {
                      "detail": "The resource does not exist, or exists but is not visible to this principal. Both answer the same way so probing reveals nothing.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 404,
                      "title": "Not found",
                      "type": "/v1/errors/not_found"
                    }
                  }
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "validation_failed": {
                    "summary": "Request validation failed",
                    "value": {
                      "detail": "A parameter failed validation, including declared caps. `errors[]` names each offending pointer. Caps are refused, never silently clamped.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Request validation failed",
                      "type": "/v1/errors/validation_failed"
                    }
                  }
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "One well header",
        "tags": [
          "wells"
        ],
        "x-glasswell-request-example": {
          "path": {
            "api10": "3305310451"
          },
          "query": {}
        }
      }
    },
...
  }
}